--- EPlast/EPlast.WebApi/StartupExtensions/AddHangFireExtension.cs
@@ -1,23 +1,0 @@
-﻿using Hangfire;
-using Hangfire.MemoryStorage;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace EPlast.WebApi.StartupExtensions
-{
-    public static class AddHangFireExtension
-    {
-        public static IServiceCollection AddHangFire(this IServiceCollection services)
-        {
-            services.AddHangfire(config =>
-            {
-                config.SetDataCompatibilityLevel(CompatibilityLevel.Version_170)
-                .UseSimpleAssemblyNameTypeSerializer()
-                .UseDefaultTypeSerializer()
-                .UseMemoryStorage();
-            }
-            );
-
-            return services;
-        }
-    }
-}
