--- EPlast/EPlast.WebApi/StartupExtensions/AddIdentityOptionsExtension.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace EPlast.WebApi.StartupExtensions
-{
-    public static class AddIdentityOptionsExtension
-    {
-        public static IServiceCollection AddIdentityOptions(this IServiceCollection services)
-        {
-            services.Configure<IdentityOptions>(options =>
-            {
-                options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(15);
-                options.Lockout.MaxFailedAccessAttempts = 5;
-                options.Password.RequiredLength = 8;
-                options.SignIn.RequireConfirmedEmail = true;
-                options.User.RequireUniqueEmail = true;
-            });
-
-            return services;
-        }
-    }
-}
